Major change
Adding a custom converter between CMat (custom matrix) and numpy ndarray.
Using this converter, functions can have CMat as in and output signatures, while automatically being transformed from and to a numpy array in the python code.
This completely hides the binding from users on C++ and python side.

The implementation was done for a double CMat and py::array_t but it is very easy to change to a CMat of other type (or to a cv::Mat_)

The returned numpy array is based on the allocated opencv matrix data. This means that the owndata flag in the numpy array is set to false. This is a behavior to be changed.
In contrary to existing converters, we only do the minimum number of allocations however, which is an important point for efficiency of big arrays.

The memory model of handing back an array from C++ to numpy, is assured use a pybind11::capsule that frees the cv::Mat data when the python object is deleted. In order to ensure this a hack has been made: We allocate a new cv::Mat array, which is quiet cheap because of the implicit sharing and we use its internal mechanism of reference countaint, for a destruction when necessary of the pypbind capsule (se file matrix-convert.inl.hh)