Fixing bug that converter was specialized only on double type but can easily be written as a template for all types
Fixing important bug for memory model: Returning an input array, won't lead to segmentation fault anymore because a deep copy is automatically made on opencv::Matrix objects that are non owning (defered by the u-pointer). This ensures that we always return a valid object with its own lifetime memory management.

For the case of a pass through function, that just returns the input numpy array, the disadvantage is obvious: We have to perform a deep copy because we can't reference the original (input) numpy array.